10 - Airflow Tutorial/requirements.txt
apache-airflow
apache-airflow-providers-postgres
pandas
elasticsearch